Add docstrings to fixtures
