+++ tests/test_cell_statistics.py
@@ -5,45 +5,7 @@ from unittest.mock import patch
 import numpy as np
 import pytest
 
-import napari
-
-import epitools._sample_data
 from epitools.analysis import calculate_cell_statistics
-
-
-@pytest.fixture(scope="function")
-def projected_image() -> napari.layers.Image:
-    data, kwargs, layer_type = epitools._sample_data.load_projected_data()[0]
-    return napari.layers.Image(data, **kwargs)
-
-
-@pytest.fixture(scope="function")
-def seeds_and_labels(
-    make_napari_viewer,
-) -> tuple[napari.layers.Points, napari.layers.Labels]:
-    seeds_layer_data, labels_layer_data = epitools._sample_data.load_segmented_data()
-    seeds_data, seeds_kwargs, _ = seeds_layer_data
-    labels_data, labels_kwargs, _ = labels_layer_data
-    return (
-        napari.layers.Points(seeds_data, **seeds_kwargs),
-        napari.layers.Labels(labels_data, **labels_kwargs),
-    )
-
-
-@pytest.fixture(scope="function")
-def viewer_with_image(
-    make_napari_viewer,
-    projected_image,
-    seeds_and_labels,
-) -> napari.Viewer:
-    seeds, labels = seeds_and_labels
-    viewer = make_napari_viewer()
-
-    viewer.add_layer(projected_image)
-    viewer.add_layer(labels)
-    viewer.add_layer(seeds)
-
-    return viewer
 
 
 def test_add_cell_statistics_widget(make_napari_viewer):
@@ -60,7 +22,7 @@ def test_add_cell_statistics_widget(make_napari_viewer):
 
 
 def test_segmentation_widget_run_button(
-    viewer_with_image,
+    viewer_with_segmentation,
     seeds_and_labels,
 ):
     """
@@ -69,7 +31,7 @@ def test_segmentation_widget_run_button(
     has it's features set to the calculated cell statistics.
     """
 
-    dock_widget, container = viewer_with_image.window.add_plugin_dock_widget(
+    dock_widget, container = viewer_with_segmentation.window.add_plugin_dock_widget(
         plugin_name="epitools",
         widget_name="Cell statistics",
     )
@@ -79,8 +41,8 @@ def test_segmentation_widget_run_button(
 
     # Check cells layer has no features (i.e. no cell stats)
     # before pressing the 'Run' button
-    image_layer, cells_layer, seeds_layer = viewer_with_image.layers
-    assert not viewer_with_image.layers["Cells"].features.size
+    image_layer, cells_layer, seeds_layer = viewer_with_segmentation.layers
+    assert not viewer_with_segmentation.layers["Cells"].features.size
 
     # use saved image data so we don't run the segmentation analysis
     # when the button is pressed
